server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
argon2-cffi==23.1.0
email-validator==2.2.0
python-multipart==0.0.12
